Some progress in host cleanup
Still lots of improvements needed.
