+++ postform-host/Cargo.toml
@@ -11,4 +11,4 @@ structopt = "0.3"
 object = "0.22.0"
 termion = "1.5.5"
 cobs = "0.1.4"
-
+strum_macros = "0.20"
